requirements.txt
streamlit
pandas
plotly
Pillow
openpyxl
xlsxwriter